add LongListConverter class

+++ app/src/main/java/tremend/com/moviedb/data/vo/Movie.kt
@@ -7,6 +7,6 @@ import com.google.gson.annotations.SerializedName
 @Entity(tableName = "movies")
 data class Movie(@PrimaryKey val id: Long, val title: String, val popularity: Float,
                  @SerializedName("original_language") val originalLanguage: String?,
-                 @SerializedName("genre_ids") val genreIds: List<Int>,
+                 @SerializedName("genre_ids") val genreIds: List<Long>,
                  @SerializedName("original_title") val originalTitle: String?,
                  val adult: Boolean, val overview: String?, val releaseDate: String?)
